v2.0.6: Bug fix, styles, Clean redundant data, lower actions needed
