+++ src/route/app/(container)/navbar/index.tsx
@@ -18,7 +18,7 @@ export default function Navbar() {
     'bg-transparent fill-slate-200 shadow-none w-full hover:bg-slate-700/75 rounded-none';
 
   return (
-    <div className="w-24 h-full bg-slate-700 p-4 grid grid-cols-1 grid-rows-[repeat(8,_minmax(0,_1fr))] gap-4">
+    <div className="w-20 h-full bg-slate-700 p-2 grid grid-cols-1 grid-rows-[repeat(8,_minmax(0,_1fr))] gap-4">
       <div className="row-span-1" />
       <div className="row-span-6 flex justify-center items-center">
         <div className="grid grid-cols-1 grid-rows-3 gap-2 bg-slate-800 rounded-2xl shadow-lg w-full h-fit">
